cfg-diff and leafpad ProFontWindows check
cfg-diff made better to handle absence of diffmerge and show changed files
lib-check-system added maybe_file_has_text to check a file for text only if it exists
check-system checks leafpad editor for ProFontWindows font default

+++ bin/cfg-diff.sh
@@ -1,22 +1,51 @@
 #!/bin/bash
-# backup kde/thunderbird config files for comparison
+# backup home dir configuration files to allow a GUI config
+# change and then figure out which files changed and what the
+# differences were.
+
 DIR=workspace/backup/settings
-pushd $HOME
-rm -rf $DIR
+
+DIFF=`which diffmerge`
+[ -z $DIFF ] && DIFF=`which vimdiff`
+[ -z $DIFF ] && DIFF=`which colordiff`
+[ -z $DIFF ] && DIFF=diff
+if echo $DIFF | grep diffmerge ; then
+	DIFF="$DIFF --nosplash"
+	WAIT='&'
+else
+	WAIT=';'
+fi
+
+echo diff setup: $DIFF new old  $WAIT
+
+pushd $HOME > /dev/null
+
+[ -d $DIR ] && rm -rf $DIR
 mkdir -p $DIR
-cp -r .kde $DIR/kde
-cp -r .config $DIR/config
+[ -d .kde ] && cp -r .kde $DIR/kde
+[ -d .config ] && cp -r .config $DIR/config
 # perforce merge tool
-cp -r .p4merge $DIR/p4merge
+[ -d .p4merge ] && cp -r .p4merge $DIR/p4merge
 # postgres and mysql configs backed up
 cp .pg* .my* $DIR/
-#cp -r .thunderbird $DIR/thunderbird
+
+[ -d .thunderbird ] && cp -r .thunderbird $DIR/thunderbird
+
+echo "Configuration files with changes:" > reconfigure.newer
 touch reconfigure.timestamp
-echo Make some configuration changes in KDE/firefox now!
+
+ls $DIR
+echo Make some configuration changes in GUI - KDE/firefox now!
 read WAIT
-#diffmerge --nosplash .thunderbird/ $DIR/thunderbird &
+
+find . -newer reconfigure.timestamp >> reconfigure.newer
+cat reconfigure.newer | grep -v reconfigure.newer | less
+
+[ -d $DIR/thunderbird ] && $DIFF .thunderbird/ $DIR/thunderbird $WAIT
 #diffmerge --nosplash .pgadmin3 $DIR/.pgadmin3 &
-diffmerge --nosplash .p4merge/ $DIR/p4merge &
-diffmerge --nosplash .kde/ $DIR/kde &
-diffmerge --nosplash .config/ $DIR/config &
-popd
+[ -d $DIR/p4merge ] && $DIFF .p4merge/ $DIR/p4merge $WAIT
+[ -d $DIR/kde ] && $DIFF .kde/ $DIR/kde $WAIT
+[ -d $DIR/config ] && $DIFF .config/ $DIR/config $WAIT
+
+popd > /dev/null
+
